+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Version 2.4.5
 * [292](https://github.com/mlebreuil/netbox-contract/issues/292) bug fix. Contract detail show internal party value instead of label.
 
+* [294](https://github.com/mlebreuil/netbox-contract/issues/294) Add `contract_assignments_display` plugin setting (`tab`, `inline`, `both`) for contract assignment UI. The tab view allows customization of the contracts list table columns.
 
 ### Version 2.4.4
 * [288](https://github.com/mlebreuil/netbox-contract/issues/288) Add the possibility to assign contract any object type. By default the following objects types: 'circuits.circuit', 'circuits.virtualcircuit', 'dcim.site', 'dcim.device', 'dcim.rack', 'virtualization.virtualmachine', 'virtualization.cluster', 'ipam.ipaddress', 'ipam.prefix'. This list can be overriden within the PLUGINS_CONFIG configuration parameter. Check the README file or [documentation](https://mlebreuil.github.io/netbox-contract/) for more information.